# docs/swagger.yaml
basePath: /
definitions:
  addEbook.Ebook:
    properties:
      date:
        type: string
      extension:
        type: string
      id:
        type: string
      name:
        type: string
      pk:
        type: string
      size:
        type: number
      sk:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        type: string
      username:
        type: string
    type: object
  addEbook.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  addEbook.Response:
    properties:
      ebook:
        $ref: '#/definitions/addEbook.Ebook'
      message:
        type: string
    type: object
  addFile.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  addFile.Files:
    properties:
      extension:
        type: string
      filename:
        type: string
      pk:
        type: string
      sk:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        type: string
      username:
        type: string
      versions:
        items:
          $ref: '#/definitions/addFile.Version'
        type: array
    type: object
  addFile.FormData:
    type: object
  addFile.Request:
    properties:
      file:
        $ref: '#/definitions/addFile.FormData'
      filename:
        type: string
      size:
        type: number
      tags:
        items:
          type: string
        type: array
      type:
        type: string
    type: object
  addFile.Response:
    properties:
      file:
        $ref: '#/definitions/addFile.Files'
      message:
        type: string
    type: object
  addFile.Version:
    properties:
      date:
        type: string
      id:
        type: string
      is_latest:
        type: boolean
      size:
        type: number
      version:
        type: string
    type: object
  addMessage.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  addMessage.Request:
    properties:
      message:
        type: string
      size:
        type: number
    type: object
  addMessage.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  authMaster.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  authMaster.SuccessResponse:
    properties:
      data:
        type: string
      status:
        type: integer
    type: object
  authMaster.UserAuthMaster:
    properties:
      master_password:
        type: string
      token:
        type: string
    type: object
...
  authSecret.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  authSecret.SuccessResponse:
    properties:
      data:
        type: string
      status:
        type: integer
    type: object
  authSecret.UserAuthSecret:
    properties:
      secret_code:
        type: string
      token:
        type: string
    type: object
...
  createPwd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  createPwd.Params:
    properties:
      digits:
        type: boolean
      length:
        type: integer
      lowercase_letters:
        type: boolean
      special_characters:
        type: boolean
      uppercase_letters:
        type: boolean
    type: object
  createPwd.Request:
    properties:
      identifier:
        type: string
      master:
        type: string
      mode:
        type: string
      params:
        $ref: '#/definitions/createPwd.Params'
      password:
        type: string
      size:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
  createPwd.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  deleteFile.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  deleteFile.SuccessResponse:
    properties:
      message:
        type: string
    type: object
  deleteFile.UserRequest:
    properties:
      filename:
        type: string
      id:
        type: string
      type:
        type: string
    type: object
...
  deletePwd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  deletePwd.SuccessResponse:
    properties:
      message:
        type: string
    type: object
...
  exists.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  exists.SuccessResponse:
    properties:
      exists:
        type: boolean
    type: object
  exists.UserRequest:
    properties:
      filename:
        type: string
      type:
        type: string
    type: object
  getEbooks.Ebook:
    properties:
      date:
        type: string
      extension:
        type: string
      id:
        type: string
      name:
        type: string
      size:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
  getEbooks.ErrorResponse:
    properties:
      error:
        type: string
    type: object
...
  getFiles.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  getFiles.Files:
    properties:
      extension:
        type: string
      filename:
        type: string
      pk:
        type: string
      sk:
        type: string
      tags:
        items:
          type: string
        type: array
      type:
        type: string
      username:
        type: string
      versions:
        items:
          $ref: '#/definitions/getFiles.Version'
        type: array
    type: object
  getFiles.UserRequest:
    properties:
      token:
        type: string
    type: object
  getFiles.Version:
    properties:
      date:
        type: string
      id:
        type: string
      is_latest:
        type: boolean
      size:
        type: number
      version:
        type: string
    type: object
  getMessages.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  getMessages.Message:
    properties:
      date:
        type: string
      hour:
        type: string
      id:
        type: string
      message:
        type: string
    type: object
...
  getPwd.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  getPwd.Passwords:
    properties:
      hash:
        type: string
      id:
        type: string
      identifier:
        type: string
      size:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
  getPwds.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  getPwds.Passwords:
    properties:
      hash:
        type: string
      id:
        type: string
      identifier:
        type: string
      size:
        type: number
      tags:
        items:
          type: string
        type: array
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: API for personal server.
  title: grp@server
  version: "1.0"
paths:
...
  /auth/master:
    post:
      consumes:
      - application/json
      description: Validate the master_password sent by the user and return a temp
        jwt with more duration
      parameters:
      - description: User token and master_password
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/authMaster.UserAuthMaster'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authMaster.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authMaster.ErrorResponse'
      summary: Auth and verify the master_password of the user
      tags:
      - auth
...
  /auth/secret:
    post:
      consumes:
      - application/json
      description: Validate the secret_code sent by the user and return a temp jwt
        with more duration
      parameters:
      - description: User token and secret_code
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/authSecret.UserAuthSecret'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/authSecret.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/authSecret.ErrorResponse'
      summary: Auth and verify the secret_code of the user
      tags:
      - auth
  /chat:
    get:
      consumes:
      - application/json
      description: Get all messages with the header token
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/getMessages.Message'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getMessages.ErrorResponse'
      summary: Get all messages
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Add a message with the request body
      parameters:
      - description: Request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/addMessage.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/addMessage.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/addMessage.ErrorResponse'
      summary: Add a message
      tags:
      - chat
...
  /ebooks:
    get:
      consumes:
      - application/json
      description: Get all ebooks
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/getEbooks.Ebook'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getEbooks.ErrorResponse'
      summary: Get all ebooks
      tags:
      - ebook
    post:
      consumes:
      - multipart/form-data
      description: Add a ebook with the request body
      parameters:
      - description: Request body
        in: formData
        name: ebook
        required: true
        type: file
      - description: Request body
        in: formData
        name: size
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/addEbook.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/addEbook.ErrorResponse'
      summary: Add a ebook
      tags:
      - ebook
...
  /files:
    delete:
      consumes:
      - application/json
      description: Delete a file with the request body
      parameters:
      - description: Request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/deleteFile.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deleteFile.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/deleteFile.ErrorResponse'
      summary: Delete a file
      tags:
      - file
    get:
      consumes:
      - application/json
      description: Get a files with the request body
      parameters:
      - description: Request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/getFiles.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/getFiles.Files'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getFiles.ErrorResponse'
      summary: Get a files
      tags:
      - file
    post:
      consumes:
      - application/json
      description: Add a file with the request body
      parameters:
      - description: Request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/addFile.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/addFile.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/addFile.ErrorResponse'
      summary: Add a file
      tags:
      - file
  /files/exists:
    get:
      consumes:
      - application/json
      description: Check if a file exists with the request body
      parameters:
      - description: Request body
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/exists.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/exists.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/exists.ErrorResponse'
      summary: Check if a file exists
      tags:
      - file
...
  /passwords:
    get:
      consumes:
      - application/json
      description: Retrieve all passwords associated with the user, requires a username
        in the header
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/getPwds.Passwords'
            type: array
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/getPwds.ErrorResponse'
      summary: Get user's passwords
      tags:
      - passwords
    post:
      consumes:
      - application/json
      description: Create a new password with a random password, this endpoint require
        a username in the header
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/createPwd.Request'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/createPwd.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/createPwd.ErrorResponse'
      summary: Create a new password
      tags:
      - passwords
  /passwords/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a password with the id in the path, this endpoint require
        a username in the header
      parameters:
      - description: Password id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/deletePwd.SuccessResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/deletePwd.ErrorResponse'
      summary: Delete a password
      tags:
      - passwords
    get:
      consumes:
      - application/json
      description: Get a password with the id in the path, this endpoint require a
        username in the header
      parameters:
      - description: Password id
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/getPwd.Passwords'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/getPwd.ErrorResponse'
      summary: Get a password
      tags:
      - passwords
...
swagger: "2.0"